+++ package.json
@@ -51,8 +51,7 @@
         "rollup": "^3.17.2",
         "rollup-plugin-peer-deps-external": "^2.2.4",
         "rollup-plugin-terser": "^7.0.2",
-        "storybook": "^6.5.16",
-        "styled-components": "^5.3.6"
+        "storybook": "^6.5.16"
     },
     "peerDependencies": {
         "react": "^18.2.0",
